Refactored how Intcode output is handled

+++ src/test/java/nl/erends/advent/year2019/Day05Test.java
@@ -3,7 +3,7 @@ package nl.erends.advent.year2019;
 import nl.erends.advent.util.Util;
 import org.junit.Test;
 
-import static org.junit.Assert.assertEquals;
+import static nl.erends.advent.util.Assert.assertEquals;
 
 public class Day05Test {
 
@@ -11,7 +11,7 @@ public class Day05Test {
     public void day05Test1() {
         Day05 problem = new Day05();
         problem.setInput(Util.readLine(2019, 5, 2));
-        assertEquals(5044655, problem.solve1().intValue());
-        assertEquals(7408802, problem.solve2().intValue());
+        assertEquals(5044655, problem.solve1());
+        assertEquals(7408802, problem.solve2());
     }
 }
